Update reservations slice
